DGS-3246 вынес остальные интерфейсы
